[SFrame-V3] sframe: gas: libsframe: use uint16_t for num_fres of FDE
Reduce the size of the num_fres field in the Function Descriptor Entry
(FDE) from 32 bits to 16 bits.

The number of Frame Row Entries (FREs) for a single function is extremely
unlikely to exceed 65,535 in real-world scenarios. Reducing this field
saves 2 bytes per FDE, contributing to a smaller overall SFrame section size.

(BTW, these savings will be eaten up by a later commit which adds
support for text > 2 GiB by increasing an offset from int32_t to
int64_t).

Safety checks are added to the assembler to warn and skip SFrame FDE
generation if a function's FRE count exceeds UINT16_MAX.

include/
	* sframe.h (sframe_func_desc_entry_v3): Change sfde_func_num_fres
	type to uint16_t.
gas/
	* gen-sframe.c (output_sframe_funcdesc): Write 2 bytes for num_fres
	and assert it fits in uint16_t.
	(sframe_do_fde): Add check to skip FDE emission if num_fres exceeds
	UINT16_MAX.
libsframe/
	* sframe.c (sframe_fde_tbl_init): Cast num_fres to uint16_t.
	(sframe_decode_fde): Likewise.
	(sframe_encoder_write_fde): Likewise.
	* testsuite/libsframe.decode/DATA2: Update binary test data.

+++ include/sframe.h
@@ -251,7 +251,7 @@ typedef struct sframe_func_desc_entry_v3
      beginning of the SFrame Frame Row Entry sub-section.  */
   uint32_t sfde_func_start_fre_off;
   /* Number of frame row entries for the function.  */
-  uint32_t sfde_func_num_fres;
+  uint16_t sfde_func_num_fres;
   /* Additional information for stack tracing from the function:
      - 4-bits: Identify the FRE type used for the function.
      - 1-bit: Identify the FDE type of the function - mask or inc.
